requirements.txt
aiosmtplib~=3.0.2
pyyaml~=3.0.2
